Refactor Schema class
